% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{extquery}
\alias{extquery}
\alias{ext_set}
\alias{ext_reader}
\alias{ext_writer}
\title{Get and set extension specific reader and writer.}
\usage{
ext_set(ext, rdr, wtr)

ext_reader(ext)

ext_writer(ext)
}
\arguments{
\item{ext}{File extension (without the leading .).}

\item{rdr}{Function to read files with the extension.}

\item{wtr}{Function to write to files with the extension.}
}
\description{
These functions are used to set and get the readers and writers for an
extension. This enables functions such as \code{rule_read()} and \code{rule_write()} to
read and write data without an explicitly supplied function.
\itemize{
\item \code{ext_set()} sets the reader and writer for the extension
\item \code{ext_reader()} returns the reader associated with the extension.
\item \code{ext_writer()} returns the writer associated with the extension.
}
}
