% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{extrw}
\alias{extrw}
\alias{ext_read}
\alias{ext_write}
\title{Read and write data.}
\usage{
ext_read(path)

ext_write(path, content)
}
\arguments{
\item{path}{File path.}

\item{content}{Data to write.}
}
\description{
These functions read and write data based on file's extension
\itemize{
\item \code{ext_read()} reads the contents of a file.
\item \code{ext_write()} writes the content to a file.
}
}
