% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{rule_name}
\alias{rule_name}
\alias{rule_desc}
\alias{rule_deps}
\alias{rule_dep}
\alias{rule_target}
\alias{rule_action}
\title{Query Rule objects}
\usage{
rule_name(rule)

rule_desc(rule)

rule_deps(rule)

rule_dep(rule, name)

rule_target(rule)

rule_action(rule)
}
\arguments{
\item{rule}{The rule object.}

\item{name}{Rule's dependency name.}
}
\description{
These functions can be used to query the fields of rule objects.
\itemize{
\item \code{rule_name()} returns the rule name.
\item \code{rule_desc()} returns the rule description.
\item \code{rule_deps()} returns a named list of rule's dependencies.
\item \code{rule_dep()} returns rule's dependency with the given name.
\item \code{rule_target()} returns the rule's target function.
\item \code{rule_action()} returns the rule's action function.
}
}
\examples{
q <- rule("q", "Desc Q", function(self) FALSE, function(self) print("q"))
r <- rule("r", "Desc R", function(self) FALSE, function(self) print("r"), q)
rule_name(r)
rule_desc(r)
rule_deps(r)
rule_dep(r, "q")
rule_target(r)
rule_action(r)
}
