% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read}
\alias{read}
\alias{rule_read}
\alias{rule_write}
\title{Read and write data}
\usage{
rule_read(rule, ..., ext = "")

rule_write(rule, content, ..., ext = "")
}
\arguments{
\item{rule}{The rule object.}

\item{...}{character vectors, if any values are NA, the result will also be
NA. The paths follow the recycling rules used in the tibble package,
namely that only length 1 arguments are recycled.}

\item{ext}{An optional extension to append to the generated path.}

\item{content}{Content to write to file}
}
\description{
These functions first construct a file path relative to rule's data directory
and read or write content.
\itemize{
\item \code{rule_read()} reads data.
\item \code{rule_write()} writes data.
}
}
