% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{rule}
\alias{rule}
\title{Make a new rule.}
\usage{
rule(name, desc, target, action, ...)
}
\arguments{
\item{name}{Rule name.}

\item{desc}{Rule description.}

\item{target}{Rule target function.}

\item{action}{Rule action function.}

\item{...}{Rule dependencies.}
}
\description{
Make a new rule.
}
\examples{
rule("slicer", "Slice a df", function(self) FALSE,
     function(self, df, len = 10) df[1:len, ])

}
