% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{rule_cmd}
\alias{rule_cmd}
\title{Command line interface}
\usage{
rule_cmd(rule, args = commandArgs(trailingOnly = TRUE), arg_parser = NULL)
}
\arguments{
\item{rule}{Rule object.}

\item{args}{Command-line arguments.}

\item{arg_parser}{Function to parse command-line arguments.}
}
\description{
\code{rule_cmd()} constructs a commmand-line interface for the rule tree.
This interface enables the rules to be executed with arguments supplied as
command-line parameters.
}
