% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{rule_hist}
\alias{rule_hist}
\title{Read history of rule executions.}
\usage{
rule_hist(rule, version = NULL)
}
\arguments{
\item{rule}{Rule.}

\item{version}{Version specification.}
}
\description{
Read history of rule executions.
}
