% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rule_path}
\alias{rule_path}
\title{Construct path to a file or directory starting from rule's data directory}
\usage{
rule_path(rule, ..., ext = "")
}
\arguments{
\item{rule}{The rule object.}

\item{...}{character vectors, if any values are NA, the result will also be
NA. The paths follow the recycling rules used in the tibble package,
namely that only length 1 arguments are recycled.}

\item{ext}{An optional extension to append to the generated path.}
}
\description{
Construct path to a file or directory starting from rule's data directory
}
