% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{rule_web}
\alias{rule_web}
\title{A shiny based web interface for interacting with rules.}
\usage{
rule_web(rule, port = 8000)
}
\arguments{
\item{rule}{Rule.}

\item{port}{Port for starting the shiny server.}
}
\description{
A shiny based web interface for interacting with rules.
}
